Separates feeds and syndication layout
